config added and file separted
