+++ diskQuota.c
@@ -6,57 +6,7 @@
 #include <dirent.h>
 #include <time.h>
 #include <unistd.h>
-
-typedef struct config {
-  long scan_interval; 
-  long old;
-  char directory[255];
-} Config;
-
-int readConf(Config *config) {
-  /*
-  if(!access("/etc/diskQuota.conf", F_OK)){ //conf doesnt exist try to create it
-    FILE *f = fopen("/etc/diskQuota.conf","w+");  //config file
-    printf("Failed to open config file!\n");
-    printf("Using default settings.\n");
-
-    config.scan_interval = 120;
-    fputs(f, "scan_interval=120");
-
-    config.old = 86400;
-    fputs(f, "old=120");
-
-    config.directory = "/dev/null"; 
-    fputs(f, "directory=/dev/null");
-    return 0;
-  }*/
-  FILE *f = fopen("config","r");  //config file
-  if(f == NULL) {
-    printf("Failed to open config file!\n");
-    config->scan_interval = 120;
-    config->old = 86400;
-    strcpy(config->directory ,"/dev/null"); 
-    return 0;
-  } else { //read config
-    char line[1024];
-    while( fgets(line, 1024, f) != NULL){
-      char *part;
-      if(strstr(line, "scan_interval") != NULL){
-        part = strstr(line, "=");
-        part++;
-        printf("%s", part);
-      }
-      if((part = strstr(line, "old")) != NULL){
-        
-      }
-      if((part = strstr(line, "directory")) != NULL){
-        
-      }
-    }
-  }
-  fclose(f);
-  return(1);
-}
+#include "config.c"
 
 int statFile(char *directory, Config *config) {
   struct stat st;
